US-17 setProductPrice implemented (except admin auth) - to be tested

+++ src/main/java/be/kdg/sa/clients/controller/dto/ProductDto.java
@@ -3,19 +3,20 @@ package be.kdg.sa.clients.controller.dto;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public class ProductDto {
     private UUID productId;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private String description;
     private int quantity;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
-    public ProductDto(UUID productId, String name, Double price, String description, int quantity, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    public ProductDto(UUID productId, String name, BigDecimal price, String description, int quantity, LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.productId = productId;
         this.name = name;
         this.price = price;
@@ -33,7 +34,7 @@ public class ProductDto {
         return name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
@@ -61,7 +62,7 @@ public class ProductDto {
         this.name = name;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
